# packages/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: CineWeave API
  version: 0.1.0
  description: API contract seed for the Phase 1 CineWeave control plane.
servers:
  - url: http://localhost:8080
paths:
...
  /api/workflow-runs/{workflowRunId}/shots:
    get:
      operationId: listWorkflowRunShots
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/WorkflowRunId"
        - name: includePreviewUrl
          in: query
          schema:
            type: boolean
            default: false
        - name: previewExpiresSeconds
          in: query
          schema:
            type: integer
            default: 900
      responses:
        "200":
          description: Storyboard shots for a workflow run
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoryboardShotListEnvelope"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    serviceTokenAuth:
      type: http
      scheme: bearer
      bearerFormat: service-token
  parameters:
...
    WorkflowRunId:
      name: workflowRunId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    Error:
      description: Error response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
  schemas:
    Envelope:
      type: object
      required: [requestId]
      properties:
        requestId:
          type: string
        data: {}
        meta: {}
...
    ErrorEnvelope:
      type: object
      required: [requestId, error]
      properties:
        requestId:
          type: string
        error:
          $ref: "#/components/schemas/ErrorBody"
    ErrorBody:
      type: object
      required: [code, message, retryable]
      properties:
        code:
          type: string
          description: Includes PROVIDER_RATE_LIMITED, PROVIDER_CONCURRENCY_LIMITED, PROVIDER_DAILY_QUOTA_EXCEEDED, PROVIDER_MONTHLY_BUDGET_EXCEEDED, and PROVIDER_CIRCUIT_OPEN for Provider Gateway guard blocks.
        message:
          type: string
        details: {}
        retryable:
          type: boolean
...
    StoryboardShot:
      type: object
      required: [id, workflowRunId, shotIndex, shotNo, status, reviewStatus, manualOverride, staleState]
      properties:
        id:
          type: string
          format: uuid
        workflowRunId:
          type: string
          format: uuid
        scriptSceneId:
          type: string
          format: uuid
        sourceScene:
          $ref: "#/components/schemas/ShotSourceScene"
        shotIndex:
          type: integer
        shotNo:
          type: integer
        durationSeconds:
          type: number
          format: double
        visual:
          type: string
        camera:
          type: string
        motion:
          type: string
        mood:
          type: string
        imagePrompt:
          type: string
        videoPrompt:
          type: string
        imageArtifactId:
          type: string
          format: uuid
        imageMediaFileId:
          type: string
          format: uuid
        imageStorageKey:
          type: string
        imagePreviewUrl:
          type: string
          format: uri
        videoArtifactId:
          type: string
          format: uuid
        videoMediaFileId:
          type: string
          format: uuid
        videoStorageKey:
          type: string
        videoPreviewUrl:
          type: string
          format: uri
        providerAsyncTaskId:
          type: string
          format: uuid
        externalTaskId:
          type: string
        status:
          type: string
        reviewStatus:
          $ref: "#/components/schemas/ReviewStatus"
        manualOverride:
          type: boolean
        staleState:
          $ref: "#/components/schemas/StaleState"
        editedBy:
          type: string
          format: uuid
        editedAt:
          type: string
          format: date-time
    ShotSourceScene:
      type: object
      required: [id, sceneNo, title, characters]
      properties:
        id:
          type: string
          format: uuid
        sceneNo:
          type: integer
        title:
          type: string
        location:
          type: string
        characters:
          type: array
          items:
            type: string
...
    ReviewStatus:
      type: string
      enum: [pending, approved, rejected, needs_edit]
    StaleState:
      type: string
      enum: [fresh, upstream_changed, needs_regeneration]
...
    StoryboardShotListEnvelope:
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - type: object
          properties:
            data:
              type: object
              properties:
                items:
                  type: array
                  items:
                    $ref: "#/components/schemas/StoryboardShot"
...